requirements.txt
# BeautyScan - Plateforme d'Analyse Cosmétique Intelligente
# Dependencies for Django Frontend + Integrated API

# =============================================================================
# DJANGO CORE (Frontend + API)
# =============================================================================
Django==5.0.2
django-cors-headers==4.4.0
djangorestframework==3.15.2

# =============================================================================
# JWT AUTHENTICATION & SECURITY
# =============================================================================


# =============================================================================
# SWAGGER/OPENAPI DOCUMENTATION
# =============================================================================
drf-yasg==1.21.7

# =============================================================================
# HTTP & API INTEGRATION
# =============================================================================
requests==2.32.3
fastapi==0.116.1
uvicorn==0.35.0

# =============================================================================
# ENVIRONMENT & CONFIGURATION
# =============================================================================
python-decouple==3.8
python-dotenv==1.0.1

# =============================================================================
# AI & MACHINE LEARNING
# =============================================================================
openai==1.30.5
azure-core==1.35.0
azure-search-documents==11.5.3
azure-common==1.1.28

# =============================================================================
# PAYMENT PROCESSING
# =============================================================================
stripe==7.11.0

# =============================================================================
# TESTING & DEVELOPMENT
# =============================================================================
pytest==8.0.0
pytest-django==4.8.0

# =============================================================================
# SYSTEM & LOGGING
# =============================================================================
psutil==6.1.0
structlog==24.4.0

# =============================================================================
# DATABASE
# =============================================================================
# psycopg2-binary==2.9.9  # PostgreSQL - Uncomment for production

# =============================================================================
# UTILITIES
# =============================================================================
six==1.17.0
isodate==0.7.2
typing-extensions==4.15.0
certifi==2025.8.3
urllib3==2.5.0

# =============================================================================
# AUTO-INSTALLED DEPENDENCIES
# =============================================================================
asgiref==3.9.1
cffi==1.17.1
charset-normalizer==3.4.3
idna==3.10
pycparser==2.22
sqlparse==0.5.3
tzdata==2025.2

# =============================================================================
# IMAGING (required for Django ImageField)
# =============================================================================
Pillow==10.4.0
pydantic==2.11.7
pydantic-core==2.33.2